# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Seedling Desktop API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /transplant:
    get:
      operationId: listTransplant
      tags: [transplant]
      summary: List transplant data from vw_GH_TransplantDesk
      parameters:
        - name: berryId
          in: query
          schema:
            type: integer
        - name: teamId
          in: query
          schema:
            type: integer
        - name: pollinationYear
          in: query
          schema:
            type: integer
        - name: spCrosses
          in: query
          schema:
            type: boolean
        - name: progeny
          in: query
          schema:
            type: string
        - name: programId
          in: query
          schema:
            type: integer
        - name: destinationId
          in: query
          schema:
            type: integer
        - name: active
          in: query
          schema:
            type: boolean
        - name: availablePlants
          in: query
          schema:
            type: boolean
        - name: page
          in: query
          schema:
            type: integer
        - name: pageSize
          in: query
          schema:
            type: integer
        - name: sortBy
          in: query
          schema:
            type: string
        - name: sortDir
          in: query
          schema:
            type: string
            enum: [asc, desc]
      responses:
        "200":
          description: Transplant list with pagination
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransplantListResponse"
  /transplant/totals:
    get:
      operationId: getTransplantTotals
      tags: [transplant]
      summary: Get totals for transplant list
      parameters:
        - name: berryId
          in: query
          schema:
            type: integer
        - name: teamId
          in: query
          schema:
            type: integer
        - name: pollinationYear
          in: query
          schema:
            type: integer
        - name: spCrosses
          in: query
          schema:
            type: boolean
        - name: progeny
          in: query
          schema:
            type: string
        - name: programId
          in: query
          schema:
            type: integer
        - name: destinationId
          in: query
          schema:
            type: integer
        - name: active
          in: query
          schema:
            type: boolean
        - name: availablePlants
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: Transplant totals
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransplantTotals"
  /transplant/batch:
    patch:
      operationId: batchUpdateTransplant
      tags: [transplant]
      summary: Batch update transplant adjustments
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BatchTransplantUpdate"
      responses:
        "200":
          description: Update count
          content:
            application/json:
              schema:
                type: object
                properties:
                  updatedCount:
                    type: integer
...
components:
  schemas:
...
    TransplantRow:
      type: object
      properties:
        id:
          type: integer
        progeny:
          type: string
        parent1:
          type: string
        parent2:
          type: string
        d1Program:
          type: string
        d2Program:
          type: string
        destination1:
          type: string
        destination2:
          type: string
        transplantInstructions:
          type: string
        transplantsRequired:
          type: number
        plantNumTransplanted:
          type: number
        extraTransplants:
          type: number
        d1TransplantAdjustment:
          type: integer
          nullable: true
        d2TransplantAdjustment:
          type: integer
          nullable: true
        expectedDiscardPercentage:
          type: number
        totalSeedlingShipRequestCalc:
          type: number
        d1SeedlingShipRequest:
          type: number
        d2SeedlingShipRequest:
          type: number
        breederRequestedShipDest1Adjustments:
          type: number
        breederRequestedShipDest2Adjustments:
          type: number
        breederAdjustmentDate:
          type: string
        plantNumTransAlAzar:
          type: number
        plantNumTransSpineless:
          type: number
        plantNumTransSpiny:
          type: number
      required:
        - id
    TransplantListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/TransplantRow"
        total:
          type: integer
        page:
          type: integer
        pageSize:
          type: integer
      required:
        - data
        - total
        - page
        - pageSize
    TransplantTotals:
      type: object
      properties:
        transplantsRequired:
          type: number
        plantNumTransplanted:
          type: number
        extraTransplants:
          type: number
        d1TransplantAdjustment:
          type: number
        d2TransplantAdjustment:
          type: number
        expectedDiscardPercentage:
          type: number
        totalSeedlingShipRequestCalc:
          type: number
        d1SeedlingShipRequest:
          type: number
        d2SeedlingShipRequest:
          type: number
        breederRequestedShipDest1Adjustments:
          type: number
        breederRequestedShipDest2Adjustments:
          type: number
        plantNumTransAlAzar:
          type: number
        plantNumTransSpineless:
          type: number
        plantNumTransSpiny:
          type: number
    BatchTransplantUpdate:
      type: object
      properties:
        updates:
          type: array
          items:
            type: object
            properties:
              id:
                type: integer
              d1TransplantAdjustment:
                type: integer
                nullable: true
              d2TransplantAdjustment:
                type: integer
                nullable: true
            required:
              - id
      required:
        - updates
...